# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API Documentation",
        "description": "API documentation for my Laravel application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": ""
        }
    ],
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register pengguna baru",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Admin Vanessa"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "admin.vanessa@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "admin123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil register dan dapat token"
                    },
                    "400": {
                        "description": "Gagal register"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login untuk mendapatkan token",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "admin.vanessa@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "admin123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil login dan mendapatkan token"
                    },
                    "401": {
                        "description": "Email atau password salah"
                    }
                }
            }
        },
        "/api/whoami": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Cek user yang sedang login",
                "operationId": "f45ccdab2b32438160a2918f72db4f59",
                "responses": {
                    "200": {
                        "description": "Data user berhasil ditampilkan"
                    },
                    "401": {
                        "description": "Unauthorized (token tidak valid)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout / hapus token login",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Berhasil logout"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/checkout": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Checkout produk skincare",
                "description": "Endpoint untuk melakukan pembelian produk oleh customer",
                "operationId": "b4720344c84a64fae2236ca18cb0d4a6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil checkout",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Berhasil checkout"
                                        },
                                        "order": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2025-06-17 21:15:00"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2025-06-17 21:15:00"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Data tidak valid"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Ambil semua produk skincare",
                "operationId": "1bfaa78d1c2c3848ab8165c5dadcad3e",
                "responses": {
                    "200": {
                        "description": "Daftar semua produk"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Tambahkan produk skincare baru",
                "operationId": "e8185e307706e2cd84bbcf5171e2f09d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "brand",
                                    "category",
                                    "stock",
                                    "price"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Toner Glow"
                                    },
                                    "brand": {
                                        "type": "string",
                                        "example": "Skintific"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Toner"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Toner untuk kulit glowing"
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 75000
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produk ditambahkan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Lihat detail produk skincare",
                "operationId": "7ab21ddb713501f3d4c2914892c02a21",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID produk",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail produk ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update produk skincare",
                "operationId": "6a7e11900844cd85cff66e662b82bd4c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID produk",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Toner Baru"
                                    },
                                    "brand": {
                                        "type": "string",
                                        "example": "Wardah"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Essence"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Essence pencerah wajah"
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 60000
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produk berhasil diupdate"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Hapus produk skincare",
                "operationId": "e80a6ca46716acee6b47ea1ac91663c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID produk",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produk dihapus"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Orders",
            "description": "Proses checkout pembelian produk skincare"
        },
        {
            "name": "Authentication",
            "description": "Authentication"
        },
        {
            "name": "Products",
            "description": "Products"
        }
    ]
}